Fix mapping to have extensions be authoritative, not UTIs
In recent MacOS UTIs have become less reliable and Boxer has been
affected by this. Auto-detection of named folders fails because MacOS
fails to register all the associations and also when multiple ones may
exist in a given system.

This path adds explicit extension mapping to the existing UTI one
so it takes precedence. Tested with all extension kinds and
working. This indirectly fixes the demo games included which
were failing because of this.