Add support for MT32 ROMs provided in GameBox
Adds support for a folder named "MT-32 ROMs" in the gamebox
root containing valid MT-32 ROMs.

Code will first check the Application Support folder in the user's
Library and if not present will check for this location in the
gamebox root folder. If present, game will behave normally
and show MT-32 LCD texts.

This doesn't affect the display in the Application's settings.